Fix raytracer focus, which prevented raytracing
- The problem was refactored Animator, which didn't have a way to tell
if the animator was finished or not when calling update().
- Enable debugSystem on Test2DCoordinates so that the 2D grid is shown.
- Fix all compiler warnings.

+++ src/rae/animation/AnimationSystem.cpp
@@ -16,7 +16,7 @@ void AnimationTimeline::update(Time time)
 		anim.update(playheadFrames);
 	}
 
-	playheadSeconds += time.deltaTime();
+	playheadSeconds += (float)time.deltaTime();
 
 	if (isLooping)
 	{
